Support starting animation before adding to window
If the view is not on the screen (window == nil), defer adding animation
to activity indicator until it gets added to window.
This patch allows to call `startAnimating()` while setting up the view,
so that it starts animating immediately after going on-screen. Up until now,
in order to take effect, `startAnimating()` had to be called when the view
was already on-screen.